+++ ServerCommon_LIB/Utility/public/Thread/IWorkerItem.h
@@ -6,7 +6,7 @@ class ThWorkerJob;
 class IWorkerItem : public std::enable_shared_from_this<IWorkerItem> {
  public:
   virtual ~IWorkerItem() = default;
-  virtual bool Execute(ThWorkerJob* workerJob, const DWORD ioByte, const uint64_t errorCode) = 0;
+  virtual bool Execute(ThWorkerJob* workerJob, const DWORD ioByte, const DWORD errorCode) = 0;
 };
 
 using WokerPtr = std::shared_ptr<IWorkerItem>;
